requirements.txt
beautifulsoup4==4.12.2
EbookLib==0.18
gTTS==2.3.2
pdfplumber==0.9.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
python-docx==0.8.11
pyttsx3==2.90
wxPython==4.2.0
